requirements.txt
Flask==0.10.1
PyLD==0.6.6
celery==3.1.17
flower==0.7.3
html5lib==0.999
rdflib==4.1.2
rdflib-jsonld==0.2
pyspotlight==0.6.5.2
elasticsearch==1.3.0
-e git+https://github.com/avengerpenguin/FuXi.git#egg=FuXi
gunicorn==19.1.1
requests==2.5.1
eventlet==0.16.0
redis==2.10.3
